requirements.txt
streamlit
datasets
transformers
torch
pypdf2
python-docx